Eerste exceptionTests en excel met voorbeel-BSNs

+++ src/main/java/com/example/thevault/domain/model/Klant.java
@@ -73,7 +73,7 @@ public class Klant extends Gebruiker {
         return bsn;
     }
 
-    public void setBsn(int bsn) {
+    public void setBsn(long bsn) {
         this.bsn = bsn;
     }
